% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocktails-package.R
\docType{package}
\name{cocktails-package}
\alias{cocktails}
\alias{cocktails-package}
\title{cocktails: Hadley's cocktail book}
\description{
Hadley's cocktails and the R code needed to turn them into a website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hadley/cocktails}
  \item Report bugs at \url{https://github.com/hadley/cocktails/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

}
\keyword{internal}
