% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocktails-package.R
\docType{package}
\name{cocktails-package}
\alias{cocktails}
\alias{cocktails-package}
\title{cocktails: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

}
\keyword{internal}
